PhotographyHub/check_oauth.py
import os
import django

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'core.settings')
django.setup()

from allauth.socialaccount.models import SocialApp
from django.contrib.sites.models import Site

print("--- SocialApps ---")
for app in SocialApp.objects.all():
    print(f"ID: {app.id}, Provider: {app.provider}, Name: {app.name}, Client ID: {app.client_id}")
    for site in app.sites.all():
        print(f"  Linked to Site: ID {site.id}, Domain: {site.domain}")

print("\n--- Sites ---")
for site in Site.objects.all():
    print(f"ID: {site.id}, Domain: {site.domain}, Name: {site.name}")
