PhotographyHub/requirements.txt
Django>=5.0,<6.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
celery>=5.3,<6.0
redis>=5.0,<6.0
psycopg[binary]>=3.1,<4.0
geopy>=2.4,<3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-allauth>=65.15.0
requests>=2.32.5
cryptography>=44.0.0
qrcode>=7.4.2
razorpay>=1.4.1
